Update cache increment and add new cache metrics
